requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
matplotlib>=3.7.0

# Database
psycopg2-binary>=2.9.0

# Async/WebSocket
websockets>=11.0
aiohttp>=3.9.0
asyncio

# Prediction Market APIs
# Note: py-clob-client requires Python 3.9.10+
# We'll implement a custom Polymarket client if needed
# py-clob-client>=0.1.0
kalshi-python>=1.0.0

# Sports Data
# nba_api is the primary source for NBA data (actively maintained, works!)
nba-api>=1.10.0
# sportsipy is deprecated but kept as fallback
sportsipy>=0.4.0

# Web Scraping (fallbacks)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging
loguru>=0.7.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
